Delete redaudent files
