+++ src/main/java/duke/command/ListCommand.java
@@ -3,7 +3,6 @@ package duke.command;
 import duke.Storage;
 import duke.exception.DukeException;
 import duke.tool.TaskList;
-import duke.ui.Ui;
 
 /**
  * Represents a command which list all tasks in the task list.
@@ -24,7 +23,7 @@ public class ListCommand implements Command {
 
     @Override
     public String getResponse() {
-        return Ui.getTaskListString(tasks);
+        return TaskList.getTaskListString(tasks);
     }
 
 }
